fix(benchmarks): hard-gate only deterministic metrics, log throughput as advisory
The Production Verification Gate failed intermittently because it compared
absolute ops/sec against committed baselines on shared GitHub Actions runners.
Two runs of identical code on the same runner type were observed to differ by
~2x (6.7k vs 13.0k ops/sec for single-insert) — whole-run CPU-steal that no
per-run statistic (mean, median, wider iteration window) can correct, and that
a tolerance wide enough to absorb would make the gate meaningless.

Split the gate into two tiers:
- Hard-gated (fail CI): deterministic metrics that do not depend on runner
  speed — search recall@10 (fixed by the math) and memory-per-vector (heap
  growth for a fixed fixture).
- Advisory (log only): all absolute throughput (ops/sec). Measured and reported
  with drift visibility, but never fails CI.

Also extract the benchmark bodies (dataset seeding, iteration counts, the
operation under test) into src/benchmarks/benchmark-bodies.ts, imported by both
the capture script (benchmark-production.ts) and the gate (verify-benchmarks.ts).
Previously the two scripts each encoded their own copy and had drifted — e.g.
Large Dataset Single Insert seeded 500 vectors in capture but 200 in the gate —
so the gate measured a different operation than its baseline was captured from.
Sharing the bodies makes that class of drift structurally impossible.

A hardware-neutral throughput gate (same-run A/B vs the base revision) is the
correct long-term design and is tracked as follow-up.
